refactor: analyse ogc within init
