+++ tests/test_analysis/test_analysis_ogc.py
@@ -83,7 +83,7 @@ class TestOgcAnalysis:
     """Tests for OGC analysis"""
 
     async def test_analyse_ogc_disabled(self, setup_catalog, fake_check):
-        with patch("udata_hydra.data_formats.ogc.analyse.config") as mock_config:
+        with patch("udata_hydra.data_formats.ogc.config") as mock_config:
             mock_config.OGC_ANALYSIS_ENABLED = False
             check = await fake_check()
             data_format = await detect_data_format_from_check_or_catalog(check)
@@ -91,11 +91,6 @@ class TestOgcAnalysis:
 
     async def test_analyse_wfs_success(self, setup_catalog, db, fake_check):
         check = await fake_check(url="https://example.com/geoserver/wfs?SERVICE=WFS")
-        with patch("udata_hydra.data_formats.ogc.config") as mock_config:
-            mock_config.OGC_ANALYSIS_ENABLED = True
-            mock_config.OGC_FORMATS = ["wfs", "wms"]
-            data_format = await detect_data_format_from_check_or_catalog(check)
-            assert data_format is not None and issubclass(data_format, (Wfs, Wms))
 
         mock_crs_4326 = MagicMock(spec=Crs)
         mock_crs_4326.getcode.return_value = "EPSG:4326"
@@ -108,19 +103,22 @@ class TestOgcAnalysis:
         mock_wfs = MagicMock()
         mock_wfs.contents = {"test:layer": mock_layer}
         mock_wfs.getOperationByName.return_value = mock_get_feature
-        with (
-            patch("udata_hydra.data_formats.ogc.analyse.config") as mock_config,
-            patch.dict(
-                "udata_hydra.data_formats.ogc.analyse.SERVICE_MAPPING",
-                {"wfs": {"service": lambda *a, **kw: mock_wfs, "versions": ["2.0.0"]}},
-            ),
-            patch(
-                "udata_hydra.data_formats.ogc.analyse.helpers.notify_udata", new_callable=AsyncMock
-            ) as mock_notify,
-        ):
+
+        with patch("udata_hydra.data_formats.ogc.config") as mock_config:
             mock_config.OGC_ANALYSIS_ENABLED = True
             mock_config.OGC_FORMATS = ["wfs", "wms"]
-            result = await data_format.analyse(check=check)
+            data_format = await detect_data_format_from_check_or_catalog(check)
+            assert data_format is not None and issubclass(data_format, (Wfs, Wms))
+            with (
+                patch.dict(
+                    "udata_hydra.data_formats.ogc.SERVICE_MAPPING",
+                    {"wfs": {"service": lambda *a, **kw: mock_wfs, "versions": ["2.0.0"]}},
+                ),
+                patch(
+                    "udata_hydra.data_formats.ogc.helpers.notify_udata", new_callable=AsyncMock
+                ) as mock_notify,
+            ):
+                result = await data_format.analyse(check=check)
 
         expected_metadata = {
             "format": "wfs",
@@ -148,27 +146,24 @@ class TestOgcAnalysis:
             data_format = await detect_data_format_from_check_or_catalog(check)
             assert data_format is not None and issubclass(data_format, (Wfs, Wms))
 
-        with (
-            patch("udata_hydra.data_formats.ogc.analyse.config") as mock_config,
-            patch.dict(
-                "udata_hydra.data_formats.ogc.analyse.SERVICE_MAPPING",
-                {
-                    "wfs": {
-                        "service": lambda *a, **kw: (_ for _ in ()).throw(
-                            RequestsConnectionError("Connection failed")
-                        ),
-                        "versions": ["2.0.0"],
-                    }
-                },
-                clear=False,
-            ),
-            patch(
-                "udata_hydra.data_formats.ogc.analyse.helpers.notify_udata", new_callable=AsyncMock
-            ) as mock_notify,
-        ):
-            mock_config.OGC_ANALYSIS_ENABLED = True
-            mock_config.OGC_FORMATS = ["wfs", "wms"]
-            result = await data_format.analyse(check=check)
+            with (
+                patch.dict(
+                    "udata_hydra.data_formats.ogc.SERVICE_MAPPING",
+                    {
+                        "wfs": {
+                            "service": lambda *a, **kw: (_ for _ in ()).throw(
+                                RequestsConnectionError("Connection failed")
+                            ),
+                            "versions": ["2.0.0"],
+                        }
+                    },
+                    clear=False,
+                ),
+                patch(
+                    "udata_hydra.data_formats.ogc.helpers.notify_udata", new_callable=AsyncMock
+                ) as mock_notify,
+            ):
+                result = await data_format.analyse(check=check)
 
         assert result is None
 
@@ -184,11 +179,6 @@ class TestOgcAnalysis:
 
     async def test_analyse_wfs_fallback_version(self, setup_catalog, db, fake_check):
         check = await fake_check(url="https://example.com/geoserver/wfs?SERVICE=WFS")
-        with patch("udata_hydra.data_formats.ogc.config") as mock_config:
-            mock_config.OGC_ANALYSIS_ENABLED = True
-            mock_config.OGC_FORMATS = ["wfs", "wms"]
-            data_format = await detect_data_format_from_check_or_catalog(check)
-            assert data_format is not None and issubclass(data_format, (Wfs, Wms))
 
         mock_crs_4326 = MagicMock(spec=Crs)
         mock_crs_4326.getcode.return_value = "EPSG:4326"
@@ -210,25 +200,28 @@ class TestOgcAnalysis:
                 raise ServiceException("Version not supported")
             return mock_wfs
 
-        with (
-            patch("udata_hydra.data_formats.ogc.analyse.config") as mock_config,
-            patch.dict(
-                "udata_hydra.data_formats.ogc.analyse.SERVICE_MAPPING",
-                {
-                    "wfs": {
-                        "service": lambda *a, **kw: wfs_side_effect(*a, **kw),
-                        "versions": ["2.0.0", "1.1.0", "1.0.0"],
-                    }
-                },
-                clear=False,
-            ),
-            patch(
-                "udata_hydra.data_formats.ogc.analyse.helpers.notify_udata", new_callable=AsyncMock
-            ),
-        ):
+        with patch("udata_hydra.data_formats.ogc.config") as mock_config:
             mock_config.OGC_ANALYSIS_ENABLED = True
             mock_config.OGC_FORMATS = ["wfs", "wms"]
-            result = await data_format.analyse(check=check)
+            data_format = await detect_data_format_from_check_or_catalog(check)
+            assert data_format is not None and issubclass(data_format, (Wfs, Wms))
+
+            with (
+                patch.dict(
+                    "udata_hydra.data_formats.ogc.SERVICE_MAPPING",
+                    {
+                        "wfs": {
+                            "service": lambda *a, **kw: wfs_side_effect(*a, **kw),
+                            "versions": ["2.0.0", "1.1.0", "1.0.0"],
+                        }
+                    },
+                    clear=False,
+                ),
+                patch(
+                    "udata_hydra.data_formats.ogc.helpers.notify_udata", new_callable=AsyncMock
+                ),
+            ):
+                result = await data_format.analyse(check=check)
 
         assert result is not None
         # second version tested
@@ -236,30 +229,28 @@ class TestOgcAnalysis:
 
     async def test_analyse_wfs_empty_contents(self, setup_catalog, db, fake_check):
         check = await fake_check(url="https://example.com/geoserver/wfs?SERVICE=WFS")
+
+        mock_wfs = MagicMock()
+        mock_wfs.contents = {}
+        mock_wfs.getOperationByName.return_value = None
+
         with patch("udata_hydra.data_formats.ogc.config") as mock_config:
             mock_config.OGC_ANALYSIS_ENABLED = True
             mock_config.OGC_FORMATS = ["wfs", "wms"]
             data_format = await detect_data_format_from_check_or_catalog(check)
             assert data_format is not None and issubclass(data_format, (Wfs, Wms))
 
-        mock_wfs = MagicMock()
-        mock_wfs.contents = {}
-        mock_wfs.getOperationByName.return_value = None
-
-        with (
-            patch("udata_hydra.data_formats.ogc.analyse.config") as mock_config,
-            patch.dict(
-                "udata_hydra.data_formats.ogc.analyse.SERVICE_MAPPING",
-                {"wfs": {"service": lambda *a, **kw: mock_wfs, "versions": ["2.0.0"]}},
-                clear=False,
-            ),
-            patch(
-                "udata_hydra.data_formats.ogc.analyse.helpers.notify_udata", new_callable=AsyncMock
-            ),
-        ):
-            mock_config.OGC_ANALYSIS_ENABLED = True
-            mock_config.OGC_FORMATS = ["wfs", "wms"]
-            result = await data_format.analyse(check=check)
+            with (
+                patch.dict(
+                    "udata_hydra.data_formats.ogc.SERVICE_MAPPING",
+                    {"wfs": {"service": lambda *a, **kw: mock_wfs, "versions": ["2.0.0"]}},
+                    clear=False,
+                ),
+                patch(
+                    "udata_hydra.data_formats.ogc.helpers.notify_udata", new_callable=AsyncMock
+                ),
+            ):
+                result = await data_format.analyse(check=check)
 
         assert result == {
             "format": "wfs",
@@ -273,11 +264,6 @@ class TestOgcAnalysis:
         check = await fake_check(
             url="https://example.com/geoserver/wfs?SERVICE=WFS&typeName=ns:my_layer"
         )
-        with patch("udata_hydra.data_formats.ogc.config") as mock_config:
-            mock_config.OGC_ANALYSIS_ENABLED = True
-            mock_config.OGC_FORMATS = ["wfs", "wms"]
-            data_format = await detect_data_format_from_check_or_catalog(check)
-            assert data_format is not None and issubclass(data_format, (Wfs, Wms))
 
         mock_layer = MagicMock()
         mock_layer.crsOptions = []
@@ -286,20 +272,23 @@ class TestOgcAnalysis:
         mock_wfs.contents = {"ns:my_layer": mock_layer}
         mock_wfs.getOperationByName.return_value = None
 
-        with (
-            patch("udata_hydra.data_formats.ogc.analyse.config") as mock_config,
-            patch.dict(
-                "udata_hydra.data_formats.ogc.analyse.SERVICE_MAPPING",
-                {"wfs": {"service": lambda *a, **kw: mock_wfs, "versions": ["2.0.0"]}},
-                clear=False,
-            ),
-            patch(
-                "udata_hydra.data_formats.ogc.analyse.helpers.notify_udata", new_callable=AsyncMock
-            ),
-        ):
+        with patch("udata_hydra.data_formats.ogc.config") as mock_config:
             mock_config.OGC_ANALYSIS_ENABLED = True
             mock_config.OGC_FORMATS = ["wfs", "wms"]
-            result = await data_format.analyse(check=check)
+            data_format = await detect_data_format_from_check_or_catalog(check)
+            assert data_format is not None and issubclass(data_format, (Wfs, Wms))
+
+            with (
+                patch.dict(
+                    "udata_hydra.data_formats.ogc.SERVICE_MAPPING",
+                    {"wfs": {"service": lambda *a, **kw: mock_wfs, "versions": ["2.0.0"]}},
+                    clear=False,
+                ),
+                patch(
+                    "udata_hydra.data_formats.ogc.helpers.notify_udata", new_callable=AsyncMock
+                ),
+            ):
+                result = await data_format.analyse(check=check)
 
         assert result is not None
         assert result["detected_layer"] == {"name": "ns:my_layer", "default_crs": None}
@@ -311,11 +300,6 @@ class TestOgcAnalysis:
         check = await fake_check(
             url="https://example.com/geoserver/wfs?SERVICE=WFS&typeName=ns:unknown"
         )
-        with patch("udata_hydra.data_formats.ogc.config") as mock_config:
-            mock_config.OGC_ANALYSIS_ENABLED = True
-            mock_config.OGC_FORMATS = ["wfs", "wms"]
-            data_format = await detect_data_format_from_check_or_catalog(check)
-            assert data_format is not None and issubclass(data_format, (Wfs, Wms))
 
         mock_layer = MagicMock()
         mock_layer.crsOptions = []
@@ -324,20 +308,23 @@ class TestOgcAnalysis:
         mock_wfs.contents = {"ns:other_layer": mock_layer}
         mock_wfs.getOperationByName.return_value = None
 
-        with (
-            patch("udata_hydra.data_formats.ogc.analyse.config") as mock_config,
-            patch.dict(
-                "udata_hydra.data_formats.ogc.analyse.SERVICE_MAPPING",
-                {"wfs": {"service": lambda *a, **kw: mock_wfs, "versions": ["2.0.0"]}},
-                clear=False,
-            ),
-            patch(
-                "udata_hydra.data_formats.ogc.analyse.helpers.notify_udata", new_callable=AsyncMock
-            ),
-        ):
+        with patch("udata_hydra.data_formats.ogc.config") as mock_config:
             mock_config.OGC_ANALYSIS_ENABLED = True
             mock_config.OGC_FORMATS = ["wfs", "wms"]
-            result = await data_format.analyse(check=check)
+            data_format = await detect_data_format_from_check_or_catalog(check)
+            assert data_format is not None and issubclass(data_format, (Wfs, Wms))
+
+            with (
+                patch.dict(
+                    "udata_hydra.data_formats.ogc.SERVICE_MAPPING",
+                    {"wfs": {"service": lambda *a, **kw: mock_wfs, "versions": ["2.0.0"]}},
+                    clear=False,
+                ),
+                patch(
+                    "udata_hydra.data_formats.ogc.helpers.notify_udata", new_callable=AsyncMock
+                ),
+            ):
+                result = await data_format.analyse(check=check)
 
         assert result is not None
         assert result["detected_layer"] is None
@@ -349,11 +336,6 @@ class TestOgcAnalysis:
         check = await fake_check(
             url="https://example.com/geoserver/wfs?SERVICE=WFS&typeName=my_layer"
         )
-        with patch("udata_hydra.data_formats.ogc.config") as mock_config:
-            mock_config.OGC_ANALYSIS_ENABLED = True
-            mock_config.OGC_FORMATS = ["wfs", "wms"]
-            data_format = await detect_data_format_from_check_or_catalog(check)
-            assert data_format is not None and issubclass(data_format, (Wfs, Wms))
 
         mock_layer = MagicMock()
         mock_layer.crsOptions = []
@@ -361,21 +343,24 @@ class TestOgcAnalysis:
         mock_wfs = MagicMock()
         mock_wfs.contents = {"ns:my_layer": mock_layer}
         mock_wfs.getOperationByName.return_value = None
-
-        with (
-            patch("udata_hydra.data_formats.ogc.analyse.config") as mock_config,
-            patch.dict(
-                "udata_hydra.data_formats.ogc.analyse.SERVICE_MAPPING",
-                {"wfs": {"service": lambda *a, **kw: mock_wfs, "versions": ["2.0.0"]}},
-                clear=False,
-            ),
-            patch(
-                "udata_hydra.data_formats.ogc.analyse.helpers.notify_udata", new_callable=AsyncMock
-            ),
-        ):
+    
+        with patch("udata_hydra.data_formats.ogc.config") as mock_config:
             mock_config.OGC_ANALYSIS_ENABLED = True
             mock_config.OGC_FORMATS = ["wfs", "wms"]
-            result = await data_format.analyse(check=check)
+            data_format = await detect_data_format_from_check_or_catalog(check)
+            assert data_format is not None and issubclass(data_format, (Wfs, Wms))
+
+            with (
+                patch.dict(
+                    "udata_hydra.data_formats.ogc.SERVICE_MAPPING",
+                    {"wfs": {"service": lambda *a, **kw: mock_wfs, "versions": ["2.0.0"]}},
+                    clear=False,
+                ),
+                patch(
+                    "udata_hydra.data_formats.ogc.helpers.notify_udata", new_callable=AsyncMock
+                ),
+            ):
+                result = await data_format.analyse(check=check)
 
         assert result is not None
         assert result["detected_layer"] == {"name": "ns:my_layer", "default_crs": None}
@@ -387,11 +372,6 @@ class TestOgcAnalysis:
         check = await fake_check(
             url="https://example.com/geoserver/wfs?SERVICE=WFS&typeName=my_layer"
         )
-        with patch("udata_hydra.data_formats.ogc.config") as mock_config:
-            mock_config.OGC_ANALYSIS_ENABLED = True
-            mock_config.OGC_FORMATS = ["wfs", "wms"]
-            data_format = await detect_data_format_from_check_or_catalog(check)
-            assert data_format is not None and issubclass(data_format, (Wfs, Wms))
 
         mock_layer = MagicMock()
         mock_layer.crsOptions = []
@@ -400,26 +380,29 @@ class TestOgcAnalysis:
         mock_wfs.contents = {"ns1:my_layer": mock_layer, "ns2:my_layer": mock_layer}
         mock_wfs.getOperationByName.return_value = None
 
-        with (
-            patch("udata_hydra.data_formats.ogc.analyse.config") as mock_config,
-            patch.dict(
-                "udata_hydra.data_formats.ogc.analyse.SERVICE_MAPPING",
-                {"wfs": {"service": lambda *a, **kw: mock_wfs, "versions": ["2.0.0"]}},
-                clear=False,
-            ),
-            patch(
-                "udata_hydra.data_formats.ogc.analyse.helpers.notify_udata", new_callable=AsyncMock
-            ),
-        ):
+        with patch("udata_hydra.data_formats.ogc.config") as mock_config:
             mock_config.OGC_ANALYSIS_ENABLED = True
             mock_config.OGC_FORMATS = ["wfs", "wms"]
-            result = await data_format.analyse(check=check)
+            data_format = await detect_data_format_from_check_or_catalog(check)
+            assert data_format is not None and issubclass(data_format, (Wfs, Wms))
+
+            with (
+                patch.dict(
+                    "udata_hydra.data_formats.ogc.SERVICE_MAPPING",
+                    {"wfs": {"service": lambda *a, **kw: mock_wfs, "versions": ["2.0.0"]}},
+                    clear=False,
+                ),
+                patch(
+                    "udata_hydra.data_formats.ogc.helpers.notify_udata", new_callable=AsyncMock
+                ),
+            ):
+                result = await data_format.analyse(check=check)
 
         assert result is not None
         assert result["detected_layer"] is None
 
     async def test_analyse_wms_disabled(self, setup_catalog, fake_check):
-        with patch("udata_hydra.data_formats.ogc.analyse.config") as mock_config:
+        with patch("udata_hydra.data_formats.ogc.config") as mock_config:
             mock_config.OGC_ANALYSIS_ENABLED = True
             mock_config.OGC_FORMATS = ["wfs"]
             check = await fake_check()
@@ -428,11 +411,6 @@ class TestOgcAnalysis:
 
     async def test_analyse_wms_success(self, setup_catalog, db, fake_check):
         check = await fake_check(url="https://example.com/geoserver/wms?SERVICE=WMS")
-        with patch("udata_hydra.data_formats.ogc.config") as mock_config:
-            mock_config.OGC_ANALYSIS_ENABLED = True
-            mock_config.OGC_FORMATS = ["wfs", "wms"]
-            data_format = await detect_data_format_from_check_or_catalog(check)
-            assert data_format is not None and issubclass(data_format, (Wfs, Wms))
 
         mock_crs_4326 = MagicMock(spec=Crs)
         mock_crs_4326.getcode.return_value = "EPSG:4326"
@@ -442,19 +420,22 @@ class TestOgcAnalysis:
         mock_wms = MagicMock()
         mock_wms.contents = {"test:layer": mock_layer}
 
-        with (
-            patch("udata_hydra.data_formats.ogc.analyse.config") as mock_config,
-            patch.dict(
-                "udata_hydra.data_formats.ogc.analyse.SERVICE_MAPPING",
-                {"wms": {"service": lambda *a, **kw: mock_wms, "versions": ["1.3.0"]}},
-            ),
-            patch(
-                "udata_hydra.data_formats.ogc.analyse.helpers.notify_udata", new_callable=AsyncMock
-            ) as mock_notify,
-        ):
+        with patch("udata_hydra.data_formats.ogc.config") as mock_config:
             mock_config.OGC_ANALYSIS_ENABLED = True
             mock_config.OGC_FORMATS = ["wfs", "wms"]
-            result = await data_format.analyse(check=check)
+            data_format = await detect_data_format_from_check_or_catalog(check)
+            assert data_format is not None and issubclass(data_format, (Wfs, Wms))
+
+            with (
+                patch.dict(
+                    "udata_hydra.data_formats.ogc.SERVICE_MAPPING",
+                    {"wms": {"service": lambda *a, **kw: mock_wms, "versions": ["1.3.0"]}},
+                ),
+                patch(
+                    "udata_hydra.data_formats.ogc.helpers.notify_udata", new_callable=AsyncMock
+                ) as mock_notify,
+            ):
+                result = await data_format.analyse(check=check)
 
         expected_metadata = {
             "format": "wms",
@@ -476,29 +457,27 @@ class TestOgcAnalysis:
 
     async def test_analyse_wms_empty_contents(self, setup_catalog, db, fake_check):
         check = await fake_check(url="https://example.com/geoserver/wms?SERVICE=WMS")
+
+        mock_wms = MagicMock()
+        mock_wms.contents = {}
+
         with patch("udata_hydra.data_formats.ogc.config") as mock_config:
             mock_config.OGC_ANALYSIS_ENABLED = True
             mock_config.OGC_FORMATS = ["wfs", "wms"]
             data_format = await detect_data_format_from_check_or_catalog(check)
             assert data_format is not None and issubclass(data_format, (Wfs, Wms))
 
-        mock_wms = MagicMock()
-        mock_wms.contents = {}
-
-        with (
-            patch("udata_hydra.data_formats.ogc.analyse.config") as mock_config,
-            patch.dict(
-                "udata_hydra.data_formats.ogc.analyse.SERVICE_MAPPING",
-                {"wms": {"service": lambda *a, **kw: mock_wms, "versions": ["1.3.0"]}},
-                clear=False,
-            ),
-            patch(
-                "udata_hydra.data_formats.ogc.analyse.helpers.notify_udata", new_callable=AsyncMock
-            ),
-        ):
-            mock_config.OGC_ANALYSIS_ENABLED = True
-            mock_config.OGC_FORMATS = ["wfs", "wms"]
-            result = await data_format.analyse(check=check)
+            with (
+                patch.dict(
+                    "udata_hydra.data_formats.ogc.SERVICE_MAPPING",
+                    {"wms": {"service": lambda *a, **kw: mock_wms, "versions": ["1.3.0"]}},
+                    clear=False,
+                ),
+                patch(
+                    "udata_hydra.data_formats.ogc.helpers.notify_udata", new_callable=AsyncMock
+                ),
+            ):
+                result = await data_format.analyse(check=check)
 
         assert result == {
             "format": "wms",
